feat(plugin): add install command and update documentation
- Introduced `install_plugin` command to install plugins from various sources.
- Updated `add_plugin_component_cmd` to utilize the new install functionality.
- Revised CLI documentation to reflect changes in plugin installation process.
